requirements.txt
flask
opencv-python
tensorflow
speechrecognition
librosa
scipy
numpy
pillow
requests
